+++ tests/workbenches/notebooks_server/controller/upgrade/test_upgrade_elyra.py
@@ -4,8 +4,6 @@ Verifies that Elyra JupyterLab extensions and runtime configurations
 survive platform upgrades without modification.
 """
 
-import json
-import re
 from typing import Any
 
 import pytest
@@ -13,170 +11,17 @@ import structlog
 from ocp_resources.notebook import Notebook
 from ocp_resources.pod import ExecOnPodError, Pod
 
+from tests.workbenches.notebooks_server.controller.upgrade.elyra_utils import (
+    ELYRA_RUNTIMES_DIR,
+    compare_runtime_config_semantics,
+    list_runtime_configs,
+    parse_elyra_extensions,
+    read_runtime_config,
+)
 from utilities.constants import Timeout
 from utilities.general import collect_pod_information
 
 LOGGER = structlog.get_logger(name=__name__)
-
-# Path to user-created Elyra runtime configurations inside workbench container
-ELYRA_RUNTIMES_DIR = "/opt/app-root/src/.local/share/jupyter/metadata/runtimes"
-
-
-def parse_elyra_extensions(labextension_output: str) -> dict[str, dict[str, Any]]:
-    """Parse jupyter labextension list output to extract Elyra-related extensions.
-
-    Matches any extension with "elyra" in the name (case-insensitive).
-
-    Args:
-        labextension_output: Raw output from `jupyter labextension list` command
-
-    Returns:
-        Dict mapping extension names to metadata (version, enabled, status)
-
-    Example:
-        Input: "odh-elyra v1.0.0 enabled OK"
-        Output: {"odh-elyra": {"version": "1.0.0", "enabled": True, "status": "OK"}}
-    """
-    elyra_extensions = {}
-
-    for line in labextension_output.split("\n"):
-        line = line.strip()
-
-        # Skip empty lines and lines without "elyra" (case-insensitive)
-        if not line or "elyra" not in line.lower():
-            continue
-
-        # Match extension line format: name v1.2.3 enabled/disabled OK/other-status
-        match = re.match(r"^([\w@/.-]+)\s+v([\d.]+)\s+(enabled|disabled)\s+(\w+)", line)
-        if match:
-            name, version, enabled_str, status = match.groups()
-            elyra_extensions[name] = {
-                "version": version,
-                "enabled": enabled_str == "enabled",
-                "status": status,
-            }
-
-    return elyra_extensions
-
-
-def list_runtime_configs(pod: Pod, container: str) -> list[str]:
-    """List Elyra runtime configuration files in the workbench pod.
-
-    Args:
-        pod: Workbench pod instance
-        container: Name of the notebook container
-
-    Returns:
-        List of runtime config filenames (e.g., ["odh_dsp.json", "custom-runtime.json"])
-
-    Raises:
-        AssertionError: If command execution fails or directory doesn't exist
-    """
-    try:
-        output = pod.execute(
-            container=container,
-            command=["sh", "-c", f"ls {ELYRA_RUNTIMES_DIR}/*.json 2>/dev/null || true"],
-            timeout=Timeout.TIMEOUT_1MIN,
-        )
-    except ExecOnPodError as e:
-        collect_pod_information(pod)
-        raise AssertionError(
-            f"Failed to list runtime configs in '{ELYRA_RUNTIMES_DIR}' on pod '{pod.name}': {e}"
-        ) from e
-
-    if not output or not output.strip():
-        return []
-
-    filenames = []
-    for line in output.strip().split("\n"):
-        line = line.strip()
-        if line and line.endswith(".json"):
-            filenames.append(line.split("/")[-1])
-
-    return filenames
-
-
-def read_runtime_config(pod: Pod, container: str, filename: str) -> dict[str, Any]:
-    """Read and parse an Elyra runtime configuration file.
-
-    Args:
-        pod: Workbench pod instance
-        container: Name of the notebook container
-        filename: Name of the runtime config file (e.g., "odh_dsp.json")
-
-    Returns:
-        Parsed JSON content as dictionary
-
-    Raises:
-        AssertionError: If file read fails or JSON is invalid
-    """
-    file_path = f"{ELYRA_RUNTIMES_DIR}/{filename}"
-
-    try:
-        output = pod.execute(
-            container=container,
-            command=["cat", file_path],
-            timeout=Timeout.TIMEOUT_1MIN,
-        )
-    except ExecOnPodError as e:
-        collect_pod_information(pod)
-        raise AssertionError(f"Failed to read runtime config '{file_path}' on pod '{pod.name}': {e}") from e
-
-    try:
-        return json.loads(output)
-    except json.JSONDecodeError as e:
-        raise AssertionError(
-            f"Runtime config '{filename}' contains invalid JSON on pod '{pod.name}'. "
-            f"Content: {output[:200]}... Error: {e}"
-        ) from e
-
-
-def compare_runtime_config_semantics(baseline: dict[str, Any], current: dict[str, Any], filename: str) -> list[str]:
-    """Compare runtime configurations semantically, focusing on critical fields.
-
-    Args:
-        baseline: Runtime config captured before upgrade
-        current: Runtime config after upgrade
-        filename: Config filename for error messages
-
-    Returns:
-        List of difference descriptions (empty if configs match semantically)
-
-    Critical fields compared:
-        - display_name
-        - schema_name
-        - metadata.runtime_type
-        - metadata.api_endpoint
-
-    Fields ignored:
-        - Timestamps, ordering, extra fields
-    """
-    differences = []
-
-    critical_fields = ["display_name", "schema_name"]
-    for field in critical_fields:
-        baseline_value = baseline.get(field)
-        current_value = current.get(field)
-
-        if baseline_value != current_value:
-            differences.append(
-                f"Field '{field}' changed: '{baseline_value}' -> '{current_value}'"
-            )
-
-    baseline_metadata = baseline.get("metadata", {})
-    current_metadata = current.get("metadata", {})
-
-    metadata_critical_fields = ["runtime_type", "api_endpoint"]
-    for field in metadata_critical_fields:
-        baseline_value = baseline_metadata.get(field)
-        current_value = current_metadata.get(field)
-
-        if baseline_value != current_value:
-            differences.append(
-                f"metadata.{field} changed: '{baseline_value}' -> '{current_value}'"
-            )
-
-    return differences
 
 
 @pytest.mark.usefixtures("capture_notebook_baseline")
@@ -272,7 +117,7 @@ class TestPreUpgradeElyra:
             missing_keys = [key for key in required_keys if key not in runtime_config]
             assert not missing_keys, (
                 f"Runtime config '{filename}' is missing required keys: {missing_keys}. "
-                f"Config content: {json.dumps(runtime_config, indent=2)}"
+                f"Found keys: {list(runtime_config.keys())}"
             )
 
             LOGGER.info(
@@ -287,10 +132,9 @@ class TestPostUpgradeElyra:
     Steps:
         1. Load pre-upgrade baseline from ConfigMap
         2. Re-execute `jupyter labextension list`
-        3. Verify extension count == baseline (exact match, no additions/removals)
-        4. Verify all baseline extensions still present with same status
-        5. Re-read runtime config files
-        6. Compare configs semantically against baseline
+        3. Verify all baseline extensions still present (allows additions, prevents removals)
+        4. Re-read runtime config files
+        5. Compare configs semantically against baseline
     """
 
     @pytest.mark.post_upgrade
@@ -302,10 +146,11 @@ class TestPostUpgradeElyra:
     ) -> None:
         """Given Elyra extensions were installed before upgrade,
         When the upgrade completes,
-        Then the exact same Elyra extensions should still be present (1:1 match).
+        Then all baseline Elyra extensions should still be present.
 
-        Requires: Exact same extension count and names
-        Prevents: Extensions being added, removed, or disabled
+        Requires: All baseline extensions still present
+        Allows: New extensions to be added
+        Prevents: Extensions being removed or disabled
         """
         baseline_extensions_list = upgrade_notebook_baseline.get("elyra_extensions_list", [])
 
@@ -332,27 +177,23 @@ class TestPostUpgradeElyra:
         baseline_extensions = set(baseline_extensions_list)
         current_extensions_set = set(current_extensions.keys())
 
-        # Check for exact 1:1 match
-        if baseline_extensions != current_extensions_set:
-            missing = baseline_extensions - current_extensions_set
-            added = current_extensions_set - baseline_extensions
+        # Check that all baseline extensions are still present (allows additions, prevents removals)
+        missing = baseline_extensions - current_extensions_set
 
-            error_parts = []
-            if missing:
-                error_parts.append(f"Missing: {', '.join(sorted(missing))}")
-            if added:
-                error_parts.append(f"Added: {', '.join(sorted(added))}")
+        assert not missing, (
+            f"The following Elyra extensions were removed during upgrade: {', '.join(sorted(missing))}. "
+            f"Pre-upgrade: {sorted(baseline_extensions)}, "
+            f"post-upgrade: {sorted(current_extensions_set)}"
+        )
 
-            raise AssertionError(
-                f"Elyra extensions changed during upgrade. "
-                f"{' | '.join(error_parts)}. "
-                f"Pre-upgrade: {sorted(baseline_extensions)}, "
-                f"post-upgrade: {sorted(current_extensions_set)}"
-            )
+        added = current_extensions_set - baseline_extensions
+        if added:
+            LOGGER.info(f"New extensions added during upgrade: {', '.join(sorted(added))}")
 
         LOGGER.info(
-            f"Elyra extensions verified: {len(current_extensions_set)} extensions, exact 1:1 match with baseline. "
-            f"Extensions: {sorted(current_extensions_set)}"
+            f"Elyra extensions verified: {len(current_extensions_set)} extensions "
+            f"({len(baseline_extensions)} baseline + {len(added)} new). "
+            f"All baseline extensions preserved."
         )
 
     @pytest.mark.post_upgrade
@@ -397,6 +238,10 @@ class TestPostUpgradeElyra:
             f"Pre-upgrade files: {sorted(baseline_filenames)}, "
             f"post-upgrade files: {sorted(current_filenames)}"
         )
+
+        added_files = current_filenames - baseline_filenames
+        if added_files:
+            LOGGER.info(f"New runtime config files added (user-created or upgrade-generated): {', '.join(sorted(added_files))}")
 
         LOGGER.info(
             f"All {len(baseline_filenames)} baseline runtime config files still exist. "
